IP-related commands
Added aliases and functions for getting internal IP (local IP) and external IPs, as well as for parsing this information to get a geographic location. The iploc function needs to be a lot cleaner; it chops off "States" in "United States" and I'm not quite sure where it's getting the state information from yet.