script to share files via HTTP over LAN
Current implementation is barebones; haven't tested with directories yet,
just files. Custom port for sharing is supported; script assumes 80 if not given.
Future versions should make sure directories can be shared, and should also provide
the PID of the netcat process, in case the file is never transferred. As it is now,
the file will be shared indefinitely until it's accessed.