Incremental monitor-setup tweaks
These changes try to handle the edge case where an already configured
and connected external monitor was disconnected, then the
`monitor-setup` script was run. The `set -u` option caused the script to
fail in that case, because the secondary_monitor was not defined. Let's
instead handle that case specially, rather than error out.