fix: chmod -x on scripts
By default chezmoi wants to store executable scripts as 644, but with
the prefix `executable_` on the filename. Only caving to this demand
because doing so allows me to `chezmoi add ~/bin/*` and commit
any new ad-hoc scripts with a minimum of fuss. Otherwise, I have to pick
through dirty git output to find what's new.