feat: nitrogen via systemd
This sounds like a great idea, but the experiment was hugely
frustrating. First off, I was motivated to try this, because on
Antigonus, sometimes the i3-startup process wouldn't properly run
`nitrogen --restore`. After upgrading to F40, that problem seems to have
gone away, so it's a moot point now.

Still, documenting the spike. Using a graphical-session user target, I
was able to get the script to run, with a noticeable one-second delay
due to restart, but that only worked on first login: if you logged out
and logged back in, the user service wouldn't run again. Madness.