feat: use gum in fedora dist-upgrade script
Trying to make the script more durable, so I don't have to edit it every
time I want to run it. Trying out using `gum` in a bash script,
which requires nix home-manager to be configured properly in order to
use. Will I regret that dependency? We shall see in six months' time.