feat: gaming-vids script riir
Mostly doing this for the CLI robustness of clap, but it's also nice to
experiment with compiling custom tools on the fly, via nix, without
having to publish things to e.g. crates.io. Really liking that.

Notably the nix tools in-repo won't autocompile; one must rerun
"install-nix" in order to get the latest flake changes. In the future,
it'll be easy enough to wire up a hook to run that command on changes to
the `flake.*` files, but let's make sure everything works on all
platforms first.